+++ Modules/Intent.Modules.Application.MediatR.Behaviours/NugetPackages.cs
@@ -18,25 +18,25 @@ namespace Intent.Modules.Application.MediatR.Behaviours
             NugetRegistry.Register(MicrosoftExtensionsLoggingPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8"),
-                        ( >= 8, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8"),
-                        ( >= 2, >= 1) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8")
-                            .WithNugetDependency("System.Diagnostics.DiagnosticSource", "9.0.8"),
-                        ( >= 2, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Bcl.AsyncInterfaces", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8")
-                            .WithNugetDependency("System.Diagnostics.DiagnosticSource", "9.0.8"),
+                        ( >= 9, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.9"),
+                        ( >= 8, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.9"),
+                        ( >= 2, >= 1) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.9")
+                            .WithNugetDependency("System.Diagnostics.DiagnosticSource", "9.0.9"),
+                        ( >= 2, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Bcl.AsyncInterfaces", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.9")
+                            .WithNugetDependency("System.Diagnostics.DiagnosticSource", "9.0.9"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftExtensionsLoggingPackageName}'"),
                     }
                 );