+++ Modules/Intent.Modules.AspNetCore.IdentityService/NugetPackages.cs
@@ -32,12 +32,12 @@ namespace Intent.Modules.AspNetCore.IdentityService
             NugetRegistry.Register(MicrosoftAspNetCoreIdentityEntityFrameworkCorePackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Identity.Stores", "9.0.8"),
-                        ( >= 8, >= 0) => new PackageVersion("8.0.19")
-                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "8.0.19")
-                            .WithNugetDependency("Microsoft.Extensions.Identity.Stores", "8.0.19"),
+                        ( >= 9, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Identity.Stores", "9.0.9"),
+                        ( >= 8, >= 0) => new PackageVersion("8.0.20")
+                            .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "8.0.20")
+                            .WithNugetDependency("Microsoft.Extensions.Identity.Stores", "8.0.20"),
                         ( >= 7, >= 0) => new PackageVersion("7.0.20")
                             .WithNugetDependency("Microsoft.EntityFrameworkCore.Relational", "7.0.20")
                             .WithNugetDependency("Microsoft.Extensions.Identity.Stores", "7.0.20"),
@@ -57,14 +57,14 @@ namespace Intent.Modules.AspNetCore.IdentityService
             NugetRegistry.Register(MicrosoftExtensionsIdentityStoresPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Identity.Core", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8"),
-                        ( >= 2, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Identity.Core", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8"),
+                        ( >= 9, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Identity.Core", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.9"),
+                        ( >= 2, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Identity.Core", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.9"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftExtensionsIdentityStoresPackageName}'"),
                     }
                 );