+++ Modules/Intent.Modules.Aws.Common/NugetPackages.cs
@@ -18,13 +18,13 @@ namespace Intent.Modules.Aws.Common
             NugetRegistry.Register(AWSSDKExtensionsNETCoreSetupPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 8, >= 0) => new PackageVersion("4.0.2.2")
-                            .WithNugetDependency("AWSSDK.Core", "4.0.0.18")
+                        ( >= 8, >= 0) => new PackageVersion("4.0.3.1")
+                            .WithNugetDependency("AWSSDK.Core", "4.0.0.29")
                             .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "2.0.0")
                             .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "8.0.0")
                             .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "2.0.0"),
-                        ( >= 2, >= 0) => new PackageVersion("4.0.2.2")
-                            .WithNugetDependency("AWSSDK.Core", "4.0.0.18")
+                        ( >= 2, >= 0) => new PackageVersion("4.0.3.1")
+                            .WithNugetDependency("AWSSDK.Core", "4.0.0.29")
                             .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "2.0.0")
                             .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "2.0.0")
                             .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "2.0.0"),