+++ Modules/Intent.Modules.Dapr.AspNetCore.ServiceInvocation/NugetPackages.cs
@@ -24,32 +24,34 @@ namespace Intent.Modules.Dapr.AspNetCore.ServiceInvocation
             NugetRegistry.Register(DaprClientPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("1.15.4")
-                            .WithNugetDependency("Dapr.Common", "1.15.4")
-                            .WithNugetDependency("Dapr.Protos", "1.15.4")
-                            .WithNugetDependency("Google.Api.CommonProtos", "2.2.0")
-                            .WithNugetDependency("Google.Protobuf", "3.30.2")
+                        ( >= 9, >= 0) => new PackageVersion("1.16.0")
+                            .WithNugetDependency("Dapr.Common", "1.16.0")
+                            .WithNugetDependency("Dapr.Protos", "1.16.0")
+                            .WithNugetDependency("Google.Api.CommonProtos", "2.17.0")
+                            .WithNugetDependency("Google.Protobuf", "3.32.0")
                             .WithNugetDependency("Grpc.Net.Client", "2.71.0")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration", "6.0.1")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "6.0.0")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "8.0.1")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "8.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Http", "6.0.0")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "6.0.0")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "6.0.4"),
-                        ( >= 8, >= 0) => new PackageVersion("1.15.4")
-                            .WithNugetDependency("Dapr.Common", "1.15.4")
-                            .WithNugetDependency("Dapr.Protos", "1.15.4")
-                            .WithNugetDependency("Google.Api.CommonProtos", "2.2.0")
-                            .WithNugetDependency("Google.Protobuf", "3.30.2")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Http", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8"),
+                        ( >= 8, >= 0) => new PackageVersion("1.16.0")
+                            .WithNugetDependency("Dapr.Common", "1.16.0")
+                            .WithNugetDependency("Dapr.Protos", "1.16.0")
+                            .WithNugetDependency("Google.Api.CommonProtos", "2.17.0")
+                            .WithNugetDependency("Google.Protobuf", "3.32.0")
                             .WithNugetDependency("Grpc.Net.Client", "2.71.0")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration", "6.0.1")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "6.0.0")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "8.0.1")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "8.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Http", "6.0.0")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "6.0.0")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "6.0.4"),
+                            .WithNugetDependency("Microsoft.Extensions.Configuration", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Http", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8"),
                         ( >= 6, >= 0) => new PackageVersion("1.15.4")
                             .WithNugetDependency("Dapr.Common", "1.15.4")
                             .WithNugetDependency("Dapr.Protos", "1.15.4")
@@ -100,10 +102,10 @@ namespace Intent.Modules.Dapr.AspNetCore.ServiceInvocation
             NugetRegistry.Register(MicrosoftAspNetCoreWebUtilitiesPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Net.Http.Headers", "9.0.8"),
-                        ( >= 8, >= 0) => new PackageVersion("8.0.19")
-                            .WithNugetDependency("Microsoft.Net.Http.Headers", "8.0.19")
+                        ( >= 9, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Net.Http.Headers", "9.0.9"),
+                        ( >= 8, >= 0) => new PackageVersion("8.0.20")
+                            .WithNugetDependency("Microsoft.Net.Http.Headers", "8.0.20")
                             .WithNugetDependency("System.IO.Pipelines", "8.0.0"),
                         ( >= 2, >= 0) => new PackageVersion("2.3.0")
                             .WithNugetDependency("Microsoft.Net.Http.Headers", "2.3.0")
@@ -114,43 +116,43 @@ namespace Intent.Modules.Dapr.AspNetCore.ServiceInvocation
             NugetRegistry.Register(MicrosoftExtensionsHttpPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Diagnostics", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8"),
-                        ( >= 8, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Diagnostics", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8"),
-                        ( >= 2, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.8")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.8"),
+                        ( >= 9, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Diagnostics", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.9"),
+                        ( >= 8, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Diagnostics", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.9"),
+                        ( >= 2, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Configuration.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "9.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "9.0.9"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftExtensionsHttpPackageName}'"),
                     }
                 );
             NugetRegistry.Register(SystemTextJsonPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 9, >= 0) => new PackageVersion("9.0.8"),
-                        ( >= 8, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("System.IO.Pipelines", "9.0.8")
-                            .WithNugetDependency("System.Text.Encodings.Web", "9.0.8"),
-                        ( >= 2, >= 0) => new PackageVersion("9.0.8")
-                            .WithNugetDependency("Microsoft.Bcl.AsyncInterfaces", "9.0.8")
+                        ( >= 9, >= 0) => new PackageVersion("9.0.9"),
+                        ( >= 8, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("System.IO.Pipelines", "9.0.9")
+                            .WithNugetDependency("System.Text.Encodings.Web", "9.0.9"),
+                        ( >= 2, >= 0) => new PackageVersion("9.0.9")
+                            .WithNugetDependency("Microsoft.Bcl.AsyncInterfaces", "9.0.9")
                             .WithNugetDependency("System.Buffers", "4.5.1")
-                            .WithNugetDependency("System.IO.Pipelines", "9.0.8")
+                            .WithNugetDependency("System.IO.Pipelines", "9.0.9")
                             .WithNugetDependency("System.Memory", "4.5.5")
                             .WithNugetDependency("System.Runtime.CompilerServices.Unsafe", "6.0.0")
-                            .WithNugetDependency("System.Text.Encodings.Web", "9.0.8")
+                            .WithNugetDependency("System.Text.Encodings.Web", "9.0.9")
                             .WithNugetDependency("System.Threading.Tasks.Extensions", "4.5.4"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{SystemTextJsonPackageName}'"),
                     }