+++ Modules/Intent.Modules.AspNetCore.IntegrationTesting/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 2.0.10
 
+- Improvement: Updated NuGet package versions.
 - Improvement: adding using clause for `HttpRequestMessage` in `HttpCLient`s.
  
 ### Version 2.0.9