+++ Modules/Intent.Modules.AspNetCore.MultiTenancy/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.1.15
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 5.1.14
 
 - Improvement: Updated documentation added help topic.