+++ Modules/Intent.Modules.Blazor/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.4
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 1.0.3
 
 - Fixed: Add type source for `Domain Enums`.