+++ Modules/Intent.Modules.MongoDb/release-notes.md
@@ -1,3 +1,7 @@
+### Version 2.0.2
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 2.0.1
 
 - Improvement: Removed document translations, transitioned to using entities.